server/requirements.txt
ultralytics>=8.0.0
websockets>=12.0
Pillow>=10.0.0
